remove addon audiodecoder.fluidsynth
There are problems with his fluidsynth library (https://www.fluidsynth.org/), as it is constructed on a Gnome basis. Therefore, the necessary GUI libraries are not available in many embedded systems.

This removes it from the repo binary addons.
Hope to get it revised sometime so that it works without Gnome.